server/requirements.txt
fastapi
pydantic
uvicorn
apscheduler
python-multipart
httpx>=0.28,<0.29
pulp
pandas
numpy
firebase-admin
python-dotenv
